# perf: dashboard monitor downsample algorithm