Partially fixes [EMQX-12962](https://emqx.atlassian.net/browse/EMQX-12962)

Release version: v/e5.8.1

## Summary

The old downsample implementation was not recursive.
So it only does reduce once, from maximum 60480 -> 30240, then that's it.

### Test result:

When the table is full (`60480` records), the latency of `/api/v5/monitor?latest=604800` call is:

- Baseline: `2.5s` (single node)
- After downsample algorithm fix: `500ms` (single node)
- Wall-clock aligned timestamps: `260ms` (single node)
- No sorting: `230ms` (single node, and the same for two-node cluster).

### Test helper:

Evaluate below expression in `remote_console` to generate random data points.

```
D = #{dropped => 0,sent => 1,connections => 1000,topics => 0,
  persisted => 0,received => 0,live_connections => 1000,
  disconnected_durable_sessions => 0,
  subscriptions_durable => 0,transformation_succeeded => 0,
  validation_failed => 0,validation_succeeded => 0}.

emqx_dashboard_monitor:randomize(60480, D, 7*timer:hours(24)).
```

To clean data:
```
emqx_dashboard_monitor:clean(0).
```

## PR Checklist
Please convert it to a draft if any of the following conditions are not met. Reviewers may skip over until all the items are checked:

- [x] Added tests for the changes
- [ ] Added property-based tests for code which performs user input validation
- [ ] Changed lines covered in coverage report
- [x] Change log has been added to `changes/(ce|ee)/(feat|perf|fix|breaking)-<PR-id>.en.md` files
- [ ] For internal contributor: there is a jira ticket to track this change
- [ ] Created PR to [emqx-docs](https://github.com/emqx/emqx-docs) if documentation update is required, or link to a follow-up jira ticket
- [ ] Schema changes are backward compatible

## Checklist for CI (.github/workflows) changes

- [ ] If changed package build workflow, pass [this action](https://github.com/emqx/emqx/actions/workflows/build_packages.yaml) (manual trigger)
- [ ] Change log has been added to `changes/` dir for user-facing artifacts update


[EMQX-12962]: https://emqx.atlassian.net/browse/EMQX-12962?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ